Add base stores.
